services/selenium_test/requirements.txt
atomicwrites==1.4.0
colorama==0.4.3
allure-pytest==2.8.11
allure-python-commons==2.8.11
attrs==19.3.0
importlib-metadata==1.5.0
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.2
pyparsing==2.4.6
pytest==6.2.4
pytest-runner==5.2
requests==2.26.0
selenium==3.141.0
six==1.14.0
wcwidth==0.1.8
zipp==3.0.0
